# RecoHGCal/TICL: optimize calculation for comparison in HGCDoublet::areAligned