By squaring cosTheta and multiplying by the squares of the magnitudes an equivalent comparison is made without the division and square root which are costly FP ops.

```
without PR

Relarted modules:

TimeReport   0.151110     0.151110     0.151110  ticlTrackstersEM
TimeReport   0.244630     0.244630     0.244630  ticlTrackstersHAD
TimeReport   5.448030     5.448030     5.448030  ticlTrackstersMIP
TimeReport   0.489669     0.489669     0.489669  ticlTrackstersMerge
TimeReport   5.136405     5.136405     5.136405  ticlTrackstersTrk

Unrelated modules:

TimeReport   2.788299     2.788299     2.788299  unsortedOfflinePrimaryVertices
TimeReport   3.190056     3.190056     3.190056  unsortedOfflinePrimaryVertices4D
TimeReport   2.398358     2.398358     2.398358  unsortedOfflinePrimaryVertices4DnoPID


MemoryReport> Peak virtual size 16747.1 Mbytes
 Key events increasing vsize:
[5] run: 1 lumi: 1 event: 6  vsize = 12483.1 deltaVsize = 680 rss = 7707.91 delta = 1096.84
[6] run: 1 lumi: 1 event: 2  vsize = 13507.1 deltaVsize = 1024 rss = 7626.52 delta = -81.3906
[11] run: 1 lumi: 1 event: 13  vsize = 15043.1 deltaVsize = 1536 rss = 7504.27 delta = -122.246
[51] run: 1 lumi: 1 event: 54  vsize = 16107.1 deltaVsize = 320 rss = 7668.46 delta = 164.191
[58] run: 1 lumi: 1 event: 59  vsize = 16427.1 deltaVsize = 320 rss = 8162.55 delta = 494.09
[88] run: 1 lumi: 1 event: 90  vsize = 16747.1 deltaVsize = 0 rss = 7822.57 delta = 143.547
[87] run: 1 lumi: 1 event: 91  vsize = 16747.1 deltaVsize = 0 rss = 7898.27 delta = 219.254
[86] run: 1 lumi: 1 event: 87  vsize = 16747.1 deltaVsize = 320 rss = 7679.02 delta = -483.535
TimeReport> Time report complete in 3095.72 seconds
 Time Summary:
 - Min event:   74.3302
 - Max event:   191.179
 - Avg event:   117.212
 - Total loop:  3024.91
 - Total init:  70.7418
 - Total job:   3095.72
 - EventSetup Lock:   0
 - EventSetup Get:   0
 Event Throughput: 0.0330588 ev/s
 CPU Summary:
 - Total loop:  11994.8
 - Total init:  44.139
 - Total extra: 0
 - Total job:   12039
 Processing Summary:
 - Number of Events:  100
 - Number of Global Begin Lumi Calls:  1
 - Number of Global Begin Run Calls: 1
```

```
with PR

Related modules

TimeReport   0.137062     0.137062     0.137062  ticlTrackstersEM
TimeReport   0.220417     0.220417     0.220417  ticlTrackstersHAD
TimeReport   4.989291     4.989291     4.989291  ticlTrackstersMIP
TimeReport   0.463667     0.463667     0.463667  ticlTrackstersMerge
TimeReport   4.469128     4.469128     4.469128  ticlTrackstersTrk

Unrelated modules

TimeReport   2.601440     2.601440     2.601440  unsortedOfflinePrimaryVertices
TimeReport   2.993901     2.993901     2.993901  unsortedOfflinePrimaryVertices4D
TimeReport   2.244631     2.244631     2.244631  unsortedOfflinePrimaryVertices4DnoPID


MemoryReport> Peak virtual size 17345.9 Mbytes
 Key events increasing vsize:
[3] run: 1 lumi: 1 event: 7  vsize = 12805.9 deltaVsize = 288 rss = 6317.03 delta = -843.027
[4] run: 1 lumi: 1 event: 1  vsize = 13229.9 deltaVsize = 424 rss = 6637.75 delta = 320.719
[10] run: 1 lumi: 1 event: 10  vsize = 16045.9 deltaVsize = 2560 rss = 7548.33 delta = 910.582
[61] run: 1 lumi: 1 event: 65  vsize = 17345.9 deltaVsize = 394 rss = 7922.04 delta = 373.711
[27] run: 1 lumi: 1 event: 28  vsize = 16621.9 deltaVsize = 320 rss = 8049.02 delta = 500.691
[63] run: 1 lumi: 1 event: 68  vsize = 17345.9 deltaVsize = 0 rss = 7725.5 delta = -196.539
[62] run: 1 lumi: 1 event: 48  vsize = 17345.9 deltaVsize = 0 rss = 7909.17 delta = -12.8672
[61] run: 1 lumi: 1 event: 65  vsize = 17345.9 deltaVsize = 394 rss = 7922.04 delta = -205.59
TimeReport> Time report complete in 2889.58 seconds
 Time Summary:
 - Min event:   69.4932
 - Max event:   169.038
 - Avg event:   109.613
 - Total loop:  2820.21
 - Total init:  69.3035
 - Total job:   2889.58
 - EventSetup Lock:   0
 - EventSetup Get:   0
 Event Throughput: 0.0354584 ev/s
 CPU Summary:
 - Total loop:  11197.6
 - Total init:  36.5688
 - Total extra: 0
 - Total job:   11234.3
 Processing Summary:
 - Number of Events:  100
 - Number of Global Begin Lumi Calls:  1
 - Number of Global Begin Run Calls: 1

```