@cmsbuild 2020-07-17
A new Pull Request was created by  @gartung (Patrick Gartung) for master.

It involves the following packages:

RecoHGCal/TICL


@perrotta, @jpata, @cmsbuild, @kpedro88, @slava77 can you please review it and eventually sign? Thanks.
@felicepantaleo, @riga, @apsallid, @sobhatta, @lecriste, @hatakeyamak, @clelange this is something you requested to watch as well.
@silviodonato, @dpiparo you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>