@cmsbuild 2020-07-20
The tests are being triggered in jenkins.