@cmsbuild 2020-07-20
-1

Tested at: cc4483924d15af7843f759ee3231a37dab4222d9

CMSSW: CMSSW_11_2_X_2020-07-20-1100
SCRAM_ARCH: slc7_amd64_gcc820
You can see the results of the tests here:
https://cmssdt.cern.ch/SDT/jenkins-artifacts/pull-request-integration/PR-3ef9f6/8142/summary.html

I found follow errors while testing this PR

**Failed tests:  RelVals**

* **RelVals**:

The relvals timed out after 4 hours.
When I ran the RelVals I found an error in the following workflows:
25202.0 step5
<pre>runTheMatrix-results/25202.0_TTbar_13+TTbar_13+DIGIUP15_PU25+RECOUP15_PU25+HARVESTUP15_PU25+NANOUP15_PU25/step5_TTbar_13+TTbar_13+DIGIUP15_PU25+RECOUP15_PU25+HARVESTUP15_PU25+NANOUP15_PU25.log

</pre>