@slava77 2020-07-21
+1

for https://github.com/cms-sw/cmssw/pull/30815	cc44839 
- code changes are in line with the PR description and the follow up comments from the community ( @VinInn;  https://mattermost.web.cern.ch/cms-o-and-c/pl/9c4k99je3b8ptbzwsccu5j7cxa)
- jenkins tests pass and comparisons with the baseline show no relevant differences, as expected from a technical update
    - timing report provided with the PR description suggests about 4% speedup in ```ticlTracksters*``` modules (after scaling to provided times in unrelated modules)